backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework_simplejwt==5.3.1
python-dotenv==1.0.1
rest_framework_simplejwt==0.0.2
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv